+++ .aws-sam/build.toml
@@ -1,0 +1,12 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.b763a3bd-0c71-4ede-a14d-61b8a64ee8da]
+codeuri = "/Users/hwangseungyeon/Desktop/personal-photo-gallery/aws/lambda/user"
+runtime = "go1.x"
+architecture = "x86_64"
+handler = "register.go"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["RegisterUserFunction"]
+
+[layer_build_definitions]
