--- .env
@@ -1,3 +1,0 @@
-# AWS
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
