+++ handlers/photo_handler.go
@@ -1,3 +1,4 @@
+// TODO: gonnado delete
 package handlers
 
 import (
