+++ samconfig.toml
@@ -1,0 +1,8 @@
+version = 0.1
+[default.deploy.parameters]
+stack_name = "PersonalPhotoGalleryStack"
+resolve_s3 = true
+s3_prefix = "PersonalPhotoGalleryStack"
+region = "ap-northeast-1"
+capabilities = "CAPABILITY_IAM"
+image_repositories = []
